+++ lib/klaro/client/resource.rb
@@ -34,6 +34,8 @@ module Klaro
   end # class Client
 end # module Klaro
 require_relative 'resource/story'
+require_relative 'resource/stories'
 require_relative 'resource/dimension'
 require_relative 'resource/dimension_value'
 require_relative 'resource/dimensions'
+require_relative 'resource/board'