+++ lib/klaro/client/resource/board.rb
@@ -1,0 +1,6 @@
+module Klaro
+  class Client
+    class Board < Resource
+    end # class Dimension
+  end # class Client
+end # module Klaro