+++ lib/klaro/client/resource/stories.rb
@@ -1,0 +1,7 @@
+module Klaro
+  class Client
+    class Stories < Collection
+      item(Story)
+    end # class Stories
+  end # class Client
+end # module Klaro