+++ CHANGELOG.md
@@ -2,11 +2,18 @@
 
 * Http bumped to "~> 4.2" which yields possible broken API since Http no longer
   exists, and is replaced by HTTP.
+
 * Client#dimensions no longer accept a dimension code. Use Client#dimension
   instead.
+
 * Client#dimension (resp. Client#dimensions) now return instances of the
   Klaro::Client::Dimension (resp. Klaro::Client::Dimensions) classes, no
   longer of ruby Hash.
+
+* Client#stories is removed. Please use Client#board_stories instead.
+
+* Board stories are no longer fully loaded by default. You must request the
+  individual story to get its full specification.
 
 # 0.3.0 - 2020-01-06